requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

Flask
Werkzeug
torch
torchvision
transformers
Pillow
gunicorn
